+++ sys/net/gnrc/network_layer/sixlowpan/frag/stats/Makefile
@@ -1,0 +1,3 @@
+MODULE := gnrc_sixlowpan_frag_stats
+
+include $(RIOTBASE)/Makefile.base
